[refactor] 메일 관련 디렉토리 하나의 디렉토리로 합침
